Scripted and safer
No customization at all. ie: You don’t choose the year or the stat page
or anything, but it’s good for what it is and what it does. I also
added a sleeper so it doesn’t kill the league. And colours I ripped off
from the Stack Overflow.